+++ tests/test_tokenization_tapas.py
@@ -44,6 +44,7 @@ class TapasTokenizationTest(TokenizerTesterMixin, unittest.TestCase):
     test_rust_tokenizer = False
     space_between_special_tokens = True
     from_pretrained_filter = filter_non_english
+    test_seq2seq = False
 
     def get_table(
         self,
